#0 | Fixing admin test for user create

+++ modules/core/src/main/java/org/openlmis/core/repository/ShipmentRoleRepository.java
@@ -36,6 +36,7 @@ public class ShipmentRoleRepository {
   }
 
   public void insertShipmentRoles(final User user) {
+    if(user.getShipmentRoles() == null) return;
     shipmentRoleAssignmentMapper.deleteAllShipmentRoles(user);
 
     for (final ShipmentRoleAssignment shipmentRoleAssignment : user.getShipmentRoles()) {
